requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.2
daphne==4.1.0
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.2
django-cache-url==3.4.5
django-cors-headers==4.3.1
environs==10.3.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
marshmallow==3.20.2
oauthlib==3.2.2
packaging==23.2
psycopg==3.1.18
psycopg-binary==3.1.18
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
whitenoise==6.6.0
zope.interface==6.1
djangorestframework==3.14.0
pillow==10.2.0
django-allauth==0.61.1